requirement.txt
absl-py
accelerate
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
aiosqlite
altair
annotated-types
antlr4-python3-runtime
anyio
appdirs
arabic-reshaper
asgiref
asn1crypto
attrs
audioop-lts
backoff
bcrypt
beautifulsoup4
bitsandbytes
blinker
Brotli
bs4
build
cachetools
certifi
cffi
cfgv
charset-normalizer
chromadb
click
colorama
coloredlogs
contourpy
crewai
cryptography
cssselect2
cycler
dataclasses-json
deep-translator
diskcache
distlib
distro
dj-database-url
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
docstring_parser
duckduckgo_search
durationpy
et_xmlfile
faiss-cpu
fastapi
fastuuid
ffmpy
filelock
filetype
flatbuffers
fonttools
freetype-py
frozenlist
fsspec
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-genai
google-generativeai
googleapis-common-protos
gradio
gradio_client
greenlet
groovy
grpcio
grpcio-status
gunicorn
h11
html5lib
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
identify
idna
immutabledict
importlib_metadata
importlib_resources
instructor
Jinja2
jiter
joblib
json5
json_repair
jsonpatch
jsonpointer
jsonref
jsonschema
jsonschema-specifications
kiwisolver
kubernetes
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
langchain-ollama
langchain-openai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
litellm
lxml
Markdown
markdown-it-py
MarkupSafe
marshmallow
matplotlib
mcp
mdurl
mmh3
mpmath
multidict
mypy_extensions
narwhals
networkx
nodeenv
numpy
oauthlib
ollama
omegaconf
onnxruntime
openai
opencv-python
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
openwakeword
orjson
ormsgpack
ortools
oscrypto
overrides
packaging
pandas
pdfminer.six
pdfplumber
pillow
pipenv
platformdirs
polars
polars-runtime-32
portalocker
posthog
pre_commit
primp
propcache
proto-plus
protobuf
psutil
psycopg2-binary
pyarrow
pyasn1
pyasn1_modules
PyAudio
pybase64
pycairo
pycparser
pydantic
pydantic-settings
pydantic_core
pydeck
pydub
pygame
Pygments
pyHanko
pyhanko-certvalidator
PyJWT
PyMuPDF
pymupdf-layout
pymupdf4llm
pyparsing
pypdf
pypdfium2
PyPika
pyproject_hooks
pyreadline3
python-bidi
python-dateutil
python-docx
python-dotenv
python-multipart
pytz
pywin32
PyYAML
RapidFuzz
referencing
regex
reportlab
requests
requests-oauthlib
requests-toolbelt
rich
rlPyCairo
rpds-py
rsa
ruff
sacremoses
safehttpx
safetensors
scikit-learn
scipy
semantic-version
sentence-transformers
sentencepiece
setuptools
shellingham
silero
six
smmap
sniffio
sounddevice
soupsieve
SQLAlchemy
sqlparse
srt
sse-starlette
starlette
streamlit
svglib
sympy
tabulate
tenacity
threadpoolctl
tiktoken
tinycss2
tokenizers
toml
tomli
tomli_w
tomlkit
torch
torchaudio
torchvision
tornado
tqdm
transformers
typer
typing-inspect
typing-inspection
typing_extensions
tzdata
tzlocal
ultralytics
ultralytics-thop
uritemplate
uritools
urllib3
uuid_utils
uv
uvicorn
virtualenv
vosk
watchdog
watchfiles
webencodings
websocket-client
websockets
whitenoise
xhtml2pdf
xxhash
yarl
zipp
zstandard
